+++ config/services.php
@@ -7,5 +7,7 @@ return [
         'environment' => env('VIVA_ENVIRONMENT', 'production'),
         'client_id' => env('VIVA_CLIENT_ID'),
         'client_secret' => env('VIVA_CLIENT_SECRET'),
+        'isv_partner_id' => env('VIVA_ISV_PARTNER_ID'),
+        'isv_partner_api_key' => env('VIVA_ISV_PARTNER_API_KEY'),
     ],
 ];