+++ tests/Functional/Services/ISV/TransactionTest.php
@@ -5,6 +5,7 @@ namespace Sebdesign\VivaPayments\Test\Functional\Services\ISV;
 use GuzzleHttp\Exception\GuzzleException;
 use GuzzleHttp\Exception\RequestException;
 use Sebdesign\VivaPayments\Facades\Viva;
+use Sebdesign\VivaPayments\Requests\CreateRecurringTransaction;
 use Sebdesign\VivaPayments\Test\TestCase;
 use Sebdesign\VivaPayments\VivaException;
 
@@ -35,5 +36,37 @@ class TransactionTest extends TestCase
         } catch (RequestException $e) {
             $this->assertEquals(404, $e->getCode());
         }
+    }
+
+    /**
+     * @test
+     * @group functional
+     * @covers \Sebdesign\VivaPayments\Requests\CreateRecurringTransaction
+     *
+     * @throws GuzzleException
+     * @throws VivaException
+     *
+     * @see https://developer.vivawallet.com/isv-partner-program/payment-isv-api/#tag/Recurring-Payments/paths/~1api~1transactions~1{id}/post
+     */
+    public function it_cannot_create_a_recurring_transaction_that_does_not_exist(): void
+    {
+        $this->expectException(VivaException::class);
+        $this->expectExceptionCode(404);
+
+        Viva::withBasicAuthCredentials(
+            strval(env('VIVA_ISV_PARTNER_ID')).':'.strval(env('VIVA_MERCHANT_ID')),
+            strval(env('VIVA_ISV_PARTNER_API_KEY')),
+        );
+
+        Viva::isv()->transactions()->createRecurring(
+            '252b950e-27f2-4300-ada1-4dedd7c17904',
+            new CreateRecurringTransaction(
+                amount: 100,
+                isvAmount: 1,
+                customerTrns: 'A description of products / services that is displayed to the customer',
+                merchantTrns: 'Your merchant reference',
+                sourceCode: '6054',
+            )
+        );
     }
 }