+++ CHANGELOG.md
@@ -5,9 +5,10 @@ All notable changes to `sebdesign/laravel-viva-payments` will be documented in t
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [6.0.0-alpha.3] - 2022-11-18
 
 - Improve exceptions
+- Create recurring transactions for ISV
 
 ## [6.0.0-alpha.2] - 2022-11-10